- refactor(miner): run mining via MinerNeuron lifecycle; CLI delegates   - Mirror previous mining loop under BaseNeuron; reuse existing helpers   - No behavior change intended
